move second test to physics
